@@ -18,76 +18,74 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.1.4",
-    "@angular/cdk": "^10.2.3",
-    "@angular/common": "^10.1.4",
-    "@angular/compiler": "^10.1.4",
-    "@angular/core": "^10.1.4",
-    "@angular/forms": "^10.1.4",
-    "@angular/material": "^11.1.1",
-    "@angular/platform-browser": "^10.1.4",
-    "@angular/platform-browser-dynamic": "^10.1.4",
-    "@angular/platform-server": "^10.1.4",
-    "@angular/router": "^10.1.4",
-    "@angular/service-worker": "^10.1.4",
-    "@apollo/client": "^3.0.0",
+    "@angular/animations": "^12.0.1",
+    "@angular/cdk": "^12.0.1",
+    "@angular/common": "^12.0.1",
+    "@angular/compiler": "^12.0.1",
+    "@angular/core": "^12.0.1",
+    "@angular/forms": "^12.0.1",
+    "@angular/platform-browser": "^12.0.1",
+    "@angular/platform-browser-dynamic": "^12.0.1",
+    "@angular/platform-server": "^12.0.1",
+    "@angular/router": "^12.0.1",
+    "@angular/service-worker": "^12.0.1",
+    "@apollo/client": "^3.3.19",
     "@fortawesome/angular-fontawesome": "^0.7.0",
     "@fortawesome/fontawesome-free": "~5.6.3",
     "@fortawesome/fontawesome-svg-core": "^1.2.31",
     "@fortawesome/free-brands-svg-icons": "^5.15.0",
     "@fortawesome/free-solid-svg-icons": "^5.15.0",
-    "@nguniversal/express-engine": "^10.1.0",
+    "@nguniversal/express-engine": "^12.0.0",
+    "apollo-angular": "^2.4.0",
     "@types/chart.js": "^2.7.40",
     "angular-bootstrap-md": "^10.1.1",
     "angular-responsive-carousel": "^2.0.2",
     "animate.css": "~3.7.2",
-    "apollo-angular": "^2.0.4",
     "bootstrap": "^4.5.2",
     "chart.js": "^2.7.2",
     "compression": "^1.7.4",
     "core-js": "^3.6.5",
     "express": "^4.15.2",
-    "graphql": "^15.0.0",
+    "graphql": "^15.5.0",
     "hammerjs": "~2.0.8",
     "photoswipe": "^4.1.3",
     "rxjs": "~6.6.3",
     "sweetalert2": "^10.15.6",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.11.1"
+    "tslib": "^2.2.0",
+    "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1001.4",
-    "@angular-devkit/build-ng-packagr": "~0.1001.4",
-    "@angular/cli": "^10.1.4",
-    "@angular/compiler-cli": "^10.1.4",
-    "@angular/language-service": "^10.1.4",
-    "@nguniversal/builders": "^10.1.0",
-    "@types/compression": "^1.7.0",
-    "@types/express": "^4.17.8",
-    "@types/graphql": "^14.0.5",
-    "@types/jasmine": "~3.5.14",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "^14.11.2",
-    "@types/photoswipe": "^4.1.0",
-    "codelyzer": "^6.0.1",
+    "@angular-devkit/build-angular": "~12.0.1",
+    "@angular/cli": "^12.0.1",
+    "@angular/compiler-cli": "^12.0.1",
+    "@angular/language-service": "^12.0.1",
     "@graphql-codegen/add": "^2.0.1",
     "@graphql-codegen/cli": "^1.13.2",
     "@graphql-codegen/fragment-matcher": "^1.17.8",
     "@graphql-codegen/typescript": "^1.13.2",
     "@graphql-codegen/typescript-compatibility": "^2.0.0",
     "@graphql-codegen/typescript-operations": "^1.13.2",
-    "jasmine-core": "~3.5.0",
+    "@nguniversal/builders": "^12.0.0",
+    "@types/compression": "^1.7.0",
+    "@types/express": "^4.17.8",
+    "@types/graphql": "^14.0.5",
+    "@types/jasmine": "~3.6.0",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "^15.3.1",
+    "@types/photoswipe": "^4.1.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.3.2",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
-    "ng-packagr": "^10.1.0",
+    "ng-packagr": "^12.0.0",
     "ts-loader": "^8.0.4",
     "ts-node": "~9.0.0",
     "tslint": "~6.1.0",
-    "typescript": "~4.0.3",
+    "typescript": "~4.2.4",
     "webpack-bundle-analyzer": "^3.9.0",
     "webpack-cli": "^3.3.12"
   },
