@@ -1,11 +1,13 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { Observable } from 'rxjs';
 
-import { AddPayment } from '../../../common/generated-types';
+import { AddPayment, GetEligiblePaymentMethods } from '../../../common/generated-types';
 import { DataService } from '../../../core/providers/data/data.service';
 import { StateService } from '../../../core/providers/state/state.service';
 
-import { ADD_PAYMENT } from './checkout-payment.graphql';
+import { ADD_PAYMENT, GET_ELIGIBLE_PAYMENT_METHODS } from './checkout-payment.graphql';
+import { map } from 'rxjs/operators';
 
 @Component({
     selector: 'vsf-checkout-payment',
@@ -13,16 +15,22 @@ import { ADD_PAYMENT } from './checkout-payment.graphql';
     styleUrls: ['./checkout-payment.component.scss'],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class CheckoutPaymentComponent {
+export class CheckoutPaymentComponent implements OnInit {
     cardNumber: string;
     expMonth: number;
     expYear: number;
+    paymentMethods$: Observable<GetEligiblePaymentMethods.EligiblePaymentMethods[]>
     paymentErrorMessage: string | undefined;
 
     constructor(private dataService: DataService,
                 private stateService: StateService,
                 private router: Router,
                 private route: ActivatedRoute) { }
+
+    ngOnInit() {
+        this.paymentMethods$ = this.dataService.query<GetEligiblePaymentMethods.Query>(GET_ELIGIBLE_PAYMENT_METHODS)
+            .pipe(map(res => res.eligiblePaymentMethods));
+    }
 
     getMonths(): number[] {
         return Array.from({ length: 12 }).map((_, i) => i + 1);
@@ -33,13 +41,11 @@ export class CheckoutPaymentComponent {
         return Array.from({ length: 10 }).map((_, i) => year + i);
     }
 
-    completeOrder() {
+    completeOrder(paymentMethodCode: string) {
         this.dataService.mutate<AddPayment.Mutation, AddPayment.Variables>(ADD_PAYMENT, {
             input: {
-                method: 'example-payment-provider',
-                metadata: {
-                    foo: 'bar',
-                },
+                method: paymentMethodCode,
+                metadata: {},
             },
         })
             .subscribe(async ({ addPaymentToOrder }) => {
@@ -47,7 +53,7 @@ export class CheckoutPaymentComponent {
                     case 'Order':
                         const order = addPaymentToOrder;
                         if (order && (order.state === 'PaymentSettled' || order.state === 'PaymentAuthorized')) {
-                            await new Promise(resolve => setTimeout(() => {
+                            await new Promise<void>(resolve => setTimeout(() => {
                                 this.stateService.setState('activeOrderId', null);
                                 resolve();
                             }, 500));
