@@ -19,21 +19,27 @@ export const SET_SHIPPING_ADDRESS = gql`
     mutation SetShippingAddress($input: CreateAddressInput!) {
         setOrderShippingAddress(input: $input) {
             ...Cart
-            shippingAddress {
-                ...OrderAddress
+            ...on Order {
+                shippingAddress {
+                    ...OrderAddress
+                }
             }
+            ...ErrorResult
         }
     }
     ${CART_FRAGMENT}
     ${ORDER_ADDRESS_FRAGMENT}
+    ${ERROR_RESULT_FRAGMENT}
 `;
 
 export const GET_ELIGIBLE_SHIPPING_METHODS = gql`
     query GetEligibleShippingMethods {
         eligibleShippingMethods {
             id
+            name
             description
             price
+            priceWithTax
             metadata
         }
     }
