@@ -110,7 +110,8 @@ export class CoreModule {
 export function apolloOptionsFactory(httpLink: HttpLink, platformId: any) {
     // Note: the intermediate assignment to `fn` is required to prevent
     // an angular compiler error. See https://stackoverflow.com/a/51977115/772859
-    let {apiHost, apiPort, shopApiPath} = environment;
+    //let {apiHost, apiPort, shopApiPath, channelToken} = environment;
+    let {apiHost, apiPort, shopApiPath } = environment;
     const isServer = isPlatformServer(platformId);
     apolloCache = new InMemoryCache({
         possibleTypes: possibleTypesData.possibleTypes,
@@ -141,6 +142,7 @@ export function apolloOptionsFactory(httpLink: HttpLink, platformId: any) {
         apiHost = process?.env?.SERVER_API_HOST || apiHost;
         apiPort = process?.env?.SERVER_API_PORT ? +process.env.SERVER_API_PORT : apiPort;
         shopApiPath = process?.env?.SERVER_API_PATH || shopApiPath;
+    //    channelToken = process?.env?.SERVER_API_CHANNEL || channelToken;
     }
     const result = {
         cache: apolloCache,
@@ -160,7 +162,9 @@ export function apolloOptionsFactory(httpLink: HttpLink, platformId: any) {
                 }
             }),
             httpLink.create({
+                //uri: `${apiHost}:${apiPort}/${shopApiPath}?vendure-token=${channelToken}`,
                 uri: `${apiHost}:${apiPort}/${shopApiPath}`,
+
                 withCredentials: true,
             })]),
     };
